requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
ta>=0.11.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
hmmlearn>=0.3.0
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
yfinance>=0.2.36
scipy>=1.11.0
websockets>=12.0
python-telegram-bot>=21.0
